crfeando componente error dialog
